ztev/server/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# server/app.py: 3
# server/config.py: 2

# server/app.py: 2
# server/config.py: 1,8
# server/resources/invite.py: 5,6
# server/resources/option.py: 5,6,11
# server/resources/role.py: 5,6
# server/resources/topic.py: 5,6,11
# server/resources/vote.py: 5,6
# server/utils.py: 1,2
Flask == 1.1.2

# server/config.py: 9
Flask_Bootstrap == 3.3.7.1

# server/config.py: 5
# server/resources/invite.py: 1,2
# server/resources/option.py: 10
# server/resources/role.py: 1,2
# server/resources/topic.py: 10
# server/resources/vote.py: 1,2
Flask_RESTful == 0.3.8

# server/config.py: 3
Flask_SQLAlchemy == 2.4.4

# server/config.py: 6,7
# server/models.py: 2,3,4,5,6
SQLAlchemy == 1.3.22

# Postgres driver. If it fails try psycopg2-binary
# On OSx refer http://stackoverflow.com/questions/22313407/clang-error-unknown-argument-mno-fused-madd-python-package-installation-fa
psycopg2-binary==2.9.4
#psycopg2-binary==2.8.6

# server/resources/topic.py: 15
python_dateutil == 2.8.1

# server/app.py: 11,12
# server/resources/invite.py: 9
# server/resources/option.py: 9
# server/resources/role.py: 9
# server/resources/topic.py: 9
# server/resources/vote.py: 9
python_jose == 3.2.0

# server/app.py: 16
requests == 2.25.1

# server/config.py: 4
sqlitedict == 1.7.0
